third_party/blink/web_tests/external/wpt/wasm/jsapi/module/moduleSource.tentative.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] AbstractModuleSource intrinsic
  Cannot access 'AbstractModuleSource' before initialization
[FAIL] AbstractModuleSourceProto intrinsic
  assert_equals: expected (object) object "[object Object]" but got (undefined) undefined
[FAIL] AbstractModuleSourceProto toStringTag brand check
  Cannot convert undefined or null to object
Harness: the test ran to completion.
